requirements_dev.txt
appnope==0.1.3; sys_platform == "darwin" and python_version >= "3.8"
asgiref==3.5.2; python_version >= "3.8"
asttokens==2.0.5; python_version >= "3.8"
backcall==0.2.0; python_version >= "3.8"
black==22.3.0; python_full_version >= "3.6.2"
certifi==2022.5.18.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3"
click==8.1.3; python_version >= "3.7" and python_full_version >= "3.6.2"
colorama==0.4.4; sys_platform == "win32" and python_version >= "3.8" and python_full_version >= "3.6.2" and platform_system == "Windows"
colorlog==6.6.0; python_version >= "3.6"
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1; python_version >= "3.8"
django-cors-headers==3.12.0; python_version >= "3.7"
django-debug-toolbar==3.4.0; python_version >= "3.7"
django-environ==0.8.1; python_version >= "3.4" and python_version < "4"
django-extensions==3.1.5; python_version >= "3.6"
django-filter==21.1; python_version >= "3.6"
django-storages==1.12.3; python_version >= "3.5"
django==4.0.5; python_version >= "3.8"
djangorestframework-simplejwt==5.2.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
executing==0.8.3; python_version >= "3.8"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
ipython==8.4.0; python_version >= "3.8"
itypes==1.2.0
jedi==0.18.1; python_version >= "3.8"
jinja2==3.1.2; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
matplotlib-inline==0.1.3; python_version >= "3.8"
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
parso==0.8.3; python_version >= "3.8"
pathspec==0.9.0; python_full_version >= "3.6.2"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.8"
pickleshare==0.7.5; python_version >= "3.8"
platformdirs==2.5.2; python_version >= "3.7" and python_full_version >= "3.6.2"
prompt-toolkit==3.0.29; python_full_version >= "3.6.2" and python_version >= "3.8"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.8"
pure-eval==0.2.2; python_version >= "3.8"
pygments==2.12.0; python_version >= "3.8"
pyjwt==2.4.0; python_version >= "3.7"
pytz==2022.1; python_version >= "3.7"
requests==2.27.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
sqlparse==0.4.2; python_version >= "3.8"
stack-data==0.2.0; python_version >= "3.8"
tomli==2.0.1; python_version < "3.11" and python_full_version >= "3.6.2" and python_version >= "3.7"
traitlets==5.2.2.post1; python_version >= "3.8"
tzdata==2022.1; sys_platform == "win32" and python_version >= "3.8"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.9; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.8"
werkzeug==2.0.2; python_version >= "3.6"
